# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AdmitGraph API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/evidence": {
      "get": {
        "tags": [
          "evidence"
        ],
        "summary": "List Evidence",
        "description": "Evidence rows (optionally one program's), scoped by the P1-10 rule.\n\nFiltering rather than 404: a collection has no single resource to refuse,\nand the caller only ever sees rows they could also read one-by-one —\nattached rows are theirs, unclaimed rows belong to the demo-mode shared\ncorpus, and anything attached to another real student never appears.",
        "operationId": "list_evidence_api_v1_evidence_get",
        "parameters": [
          {
            "name": "program_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Program Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Evidence Api V1 Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evidence/{evidence_id}": {
      "get": {
        "tags": [
          "evidence"
        ],
        "summary": "Get Evidence",
        "operationId": "get_evidence_api_v1_evidence__evidence_id__get",
        "parameters": [
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Evidence Api V1 Evidence  Evidence Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evidence/{evidence_id}/recheck": {
      "post": {
        "tags": [
          "evidence"
        ],
        "summary": "Recheck Evidence",
        "description": "Bounded fresh re-check of one claim (1 search + extraction).\n\nError order: 404 unknown/out-of-scope evidence -> 429 per-caller cooldown\n-> 429 BUDGET_EXCEEDED (daily search budget, P1-8) -> 409\nPROVIDER_UNAVAILABLE when SERPAPI_API_KEY is not configured.",
        "operationId": "recheck_evidence_api_v1_evidence__evidence_id__recheck_post",
        "parameters": [
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Recheck Evidence Api V1 Evidence  Evidence Id  Recheck Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evidence/{evidence_id}/conflicts": {
      "get": {
        "tags": [
          "evidence"
        ],
        "summary": "Get Conflicts",
        "description": "Conflict groups on one evidence row (P1-10: same gate as the read).\n\nLeft unscoped when W3 landed the access matrix — this handoff closes it:\nunknown evidence and evidence outside the caller's scope are both a 404,\nexactly like `GET /evidence/{id}`.",
        "operationId": "get_conflicts_api_v1_evidence__evidence_id__conflicts_get",
        "parameters": [
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evidence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Conflicts Api V1 Evidence  Evidence Id  Conflicts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evidence/conflicts/{conflict_id}/resolve": {
      "post": {
        "tags": [
          "evidence"
        ],
        "summary": "Resolve Conflict",
        "description": "Resolve a conflict group on the user's say-so.\n\n`preferred_evidence_id` defaults to the authority-preferred member recorded\nat detection time; only members of THIS conflict may be chosen.\n\nErrors: 404 NOT_FOUND for an unknown conflict, a conflict outside the\ncaller's scope (P1-10) or an evidence id, 409 ALREADY_RESOLVED when the\nconflict is already resolved, 409 EVIDENCE_NOT_IN_CONFLICT when the chosen\nevidence is not a member.",
        "operationId": "resolve_conflict_api_v1_evidence_conflicts__conflict_id__resolve_post",
        "parameters": [
          {
            "name": "conflict_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conflict Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ResolveConflictIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Resolve Conflict Api V1 Evidence Conflicts  Conflict Id  Resolve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolveConflictIn": {
        "properties": {
          "preferred_evidence_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Evidence Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "ResolveConflictIn",
        "description": "Optional body: resolve without a body to accept the default preference."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}